requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreschema==0.0.4
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
requests==2.31.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
